A trial version of preloading
