+++ src/scss/main.scss
@@ -1,3 +1,4 @@
+@import './preloader.scss';
 @import './base/modern-normalize';
 @import './base/common';
 @import './base/variables.scss';
